requirements.txt
youtube-transcript-api
reportlab
google-generativeai
python-dotenv
